+++ sdk/core/System.ClientModel/CHANGELOG.md
@@ -8,6 +8,8 @@
 
 ### Bugs Fixed
 
+- Fixed an issue with `ClientRetryPolicy` where delays were being calculated using the retry count instead of the attempt count, causing the initial retry to occur without delay and subsequent retries to performed more quickly than intended.
+
 ### Other Changes
 
 ## 1.8.1 (2025-11-10)